backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
qdrant-client>=1.8.0
langchain>=0.0.349
langchain-community>=0.0.349
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
aiofiles>=23.0.0
asyncpg>=0.29.0
sqlalchemy>=2.0.23
alembic>=1.13.1
pgvector>=0.2.5
openai>=1.3.5
langchain-openai>=0.0.5
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
better-auth>=0.5.1